Revert "Adding nightly run feature,  fixing CloudWatch Log spelling & adding solution architecture overview (#81)"
This reverts commit 1ee6f7dd66abfa0a9990d2c62b6f120ddf53444f.
