+++ .gitignore
@@ -21,8 +21,6 @@ cdk.out
 # Eslint config
 !.eslintrc.js
 
-# Prettier JS-doc config
-!.prettierrc
 
 # VS code specific ignores
 .vscode
