+++ bin/aws-sso-extensions-for-enterprise.ts
@@ -156,15 +156,6 @@ function getConfig() {
         unparsedEnv["Parameters"],
         "UpgradeFromVersionLessThanV303"
       ),
-      EnableNightlyRun: ensureBoolean(
-        unparsedEnv["Parameters"],
-        "EnableNightlyRun"
-      ),
-      NightlyRunRemediationMode: ensureValidString(
-        unparsedEnv["Parameters"],
-        "NightlyRunRemediationMode",
-        ["NOTIFY", "AUTOREMEDIATE"]
-      ),
     },
   };
