+++ config/env.yaml
@@ -1,7 +1,7 @@
 ---
 App: "aws-sso-extensions-for-enterprise"
 Environment: "env"
-Version: "3.2.0"
+Version: "3.1.4"
 
 PipelineSettings:
   BootstrapQualifier: "<your-bootstrap-qualifier>" # For example: 'ssoutility'
@@ -26,5 +26,3 @@ Parameters:
   DomainName: "corp.example.com" # If IsAdUsed is false, this will be ignored.
   ImportCurrentSSOConfiguration: false
   UpgradeFromVersionLessThanV303: false # Should be set to true, if upgrading from solution version 3.0.2/earlier
-  EnableNightlyRun: true
-  NightlyRunRemediationMode: "NOTIFY" # If EnableNightlyrun is set to true, this should be set to AUTOREMEDIATE or NOTIFY
