+++ lib/constructs/lambda-layers.ts
@@ -25,7 +25,7 @@ export class LambdaLayers extends Construct {
         code: Code.fromAsset(
           join(__dirname, "../", "lambda-layers", "nodejs-layer")
         ),
-        compatibleRuntimes: [Runtime.NODEJS_16_X],
+        compatibleRuntimes: [Runtime.NODEJS_14_X],
       }
     );
