+++ lib/constructs/link-crud.ts
@@ -140,7 +140,7 @@ export class LinkCRUD extends Construct {
         name(buildConfig, "linkApiHandler"),
         {
           functionName: name(buildConfig, "linkApiHandler"),
-          runtime: Runtime.NODEJS_16_X,
+          runtime: Runtime.NODEJS_14_X,
           entry: join(
             __dirname,
             "../",
@@ -192,7 +192,7 @@ export class LinkCRUD extends Construct {
         name(buildConfig, "linkCuHandler"),
         {
           functionName: name(buildConfig, "linkCuHandler"),
-          runtime: Runtime.NODEJS_16_X,
+          runtime: Runtime.NODEJS_14_X,
           entry: join(
             __dirname,
             "../",
@@ -237,7 +237,7 @@ export class LinkCRUD extends Construct {
         name(buildConfig, "linkDelHandler"),
         {
           functionName: name(buildConfig, "linkDelHandler"),
-          runtime: Runtime.NODEJS_16_X,
+          runtime: Runtime.NODEJS_14_X,
           entry: join(
             __dirname,
             "../",
