+++ lib/constructs/org-events.ts
@@ -42,7 +42,7 @@ export class OrgEvents extends Construct {
       this,
       name(buildConfig, "orgEventsHandler"),
       {
-        runtime: Runtime.NODEJS_16_X,
+        runtime: Runtime.NODEJS_14_X,
         functionName: name(buildConfig, "orgEventsHandler"),
         entry: join(
           __dirname,
