+++ lib/constructs/permission-set-processor.ts
@@ -62,7 +62,7 @@ export class PermissionSetProcessor extends Construct {
       name(buildConfig, "permissionSetTopicProcessor"),
       {
         functionName: name(buildConfig, "permissionSetTopicProcessor"),
-        runtime: lambda.Runtime.NODEJS_16_X,
+        runtime: lambda.Runtime.NODEJS_14_X,
         entry: join(
           __dirname,
           "../",
@@ -112,7 +112,7 @@ export class PermissionSetProcessor extends Construct {
       this,
       name(buildConfig, "permissionSetSyncHandler"),
       {
-        runtime: lambda.Runtime.NODEJS_16_X,
+        runtime: lambda.Runtime.NODEJS_14_X,
         functionName: name(buildConfig, "permissionSetSyncHandler"),
         entry: join(
           __dirname,
