+++ lib/stacks/pipeline/pipeline-stages.ts
@@ -12,7 +12,6 @@ import { SSOApiRoles } from "../pipelineStageStacks/sso-api-roles";
 import { SSOEventsProcessor } from "../pipelineStageStacks/sso-events-processor";
 import { SSOImportArtefactsPart1 } from "../pipelineStageStacks/sso-import-artefacts-part1";
 import { SSOImportArtefactsPart2 } from "../pipelineStageStacks/sso-import-artefacts-part2";
-import { SSOImportArtefactsPart3 } from "../pipelineStageStacks/sso-import-artefacts-part3";
 import { UpgradeToV303 } from "../pipelineStageStacks/upgrade-to-v303";
 
 function fullname(buildConfig: BuildConfig, name: string): string {
@@ -75,17 +74,19 @@ export class SSOArtefactsDeploymentStage extends Stage {
       buildConfig
     );
 
-    new SSOImportArtefactsPart1(
-      this,
-      fullname(buildConfig, "ssoImportArtefactsPart1stack"),
-      {
-        stackName: fullname(buildConfig, "ssoImportArtefactsPart1stack"),
-        synthesizer: new DefaultStackSynthesizer({
-          qualifier: buildConfig.PipelineSettings.BootstrapQualifier,
-        }),
-      },
-      buildConfig
-    );
+    if (buildConfig.Parameters.ImportCurrentSSOConfiguration) {
+      new SSOImportArtefactsPart1(
+        this,
+        fullname(buildConfig, "ssoImportArtefactsPart1stack"),
+        {
+          stackName: fullname(buildConfig, "ssoImportArtefactsPart1stack"),
+          synthesizer: new DefaultStackSynthesizer({
+            qualifier: buildConfig.PipelineSettings.BootstrapQualifier,
+          }),
+        },
+        buildConfig
+      );
+    }
   }
 }
 
@@ -124,19 +125,21 @@ export class SolutionArtefactsDeploymentStage extends Stage {
 
     solutionartefactsStack.node.addDependency(preSolutionArtefactsStack);
 
-    const ssoImportArtefactsPart2Stack = new SSOImportArtefactsPart2(
-      this,
-      fullname(buildConfig, "ssoImportArtefactsPart2Stack"),
-      {
-        stackName: fullname(buildConfig, "ssoImportArtefactsPart2Stack"),
-        synthesizer: new DefaultStackSynthesizer({
-          qualifier: buildConfig.PipelineSettings.BootstrapQualifier,
-        }),
-      },
-      buildConfig
-    );
+    if (buildConfig.Parameters.ImportCurrentSSOConfiguration) {
+      const ssoImportArtefactsPart2Stack = new SSOImportArtefactsPart2(
+        this,
+        fullname(buildConfig, "ssoImportArtefactsPart2Stack"),
+        {
+          stackName: fullname(buildConfig, "ssoImportArtefactsPart2Stack"),
+          synthesizer: new DefaultStackSynthesizer({
+            qualifier: buildConfig.PipelineSettings.BootstrapQualifier,
+          }),
+        },
+        buildConfig
+      );
 
-    ssoImportArtefactsPart2Stack.node.addDependency(solutionartefactsStack);
+      ssoImportArtefactsPart2Stack.node.addDependency(solutionartefactsStack);
+    }
 
     if (buildConfig.Parameters.UpgradeFromVersionLessThanV303) {
       const upgradeToV303Stack = new UpgradeToV303(
@@ -152,28 +155,5 @@ export class SolutionArtefactsDeploymentStage extends Stage {
       );
       upgradeToV303Stack.node.addDependency(solutionartefactsStack);
     }
-  }
-}
-
-export class SSOArtefactsPostDeployment extends Stage {
-  constructor(
-    scope: Construct,
-    id: string,
-    props: StageProps | undefined,
-    buildConfig: BuildConfig
-  ) {
-    super(scope, id, props);
-
-    new SSOImportArtefactsPart3(
-      this,
-      fullname(buildConfig, "ssoImportArtefactsPart3stack"),
-      {
-        stackName: fullname(buildConfig, "ssoImportArtefactsPart3stack"),
-        synthesizer: new DefaultStackSynthesizer({
-          qualifier: buildConfig.PipelineSettings.BootstrapQualifier,
-        }),
-      },
-      buildConfig
-    );
   }
 }
