+++ lib/stacks/pipelineStageStacks/sso-import-artefacts-part1.ts
@@ -1,20 +1,34 @@
+/*
+Deploys part 1 of artefacts required for importing
+current AWS SSO configuration i.e. permission sets and
+account assignments in SSO account
+*/
 import { Stack, StackProps } from "aws-cdk-lib";
-import { Rule, RuleTargetInput, Schedule } from "aws-cdk-lib/aws-events";
-import { SfnStateMachine } from "aws-cdk-lib/aws-events-targets";
-import { PolicyStatement } from "aws-cdk-lib/aws-iam";
-import { StateMachine } from "aws-cdk-lib/aws-stepfunctions";
+import {
+  AccountPrincipal,
+  PolicyStatement,
+  Role,
+  ServicePrincipal,
+} from "aws-cdk-lib/aws-iam";
+import { Key } from "aws-cdk-lib/aws-kms";
+import { Topic } from "aws-cdk-lib/aws-sns";
+import { CfnStateMachine } from "aws-cdk-lib/aws-stepfunctions";
 import { Construct } from "constructs";
-import { v4 as uuidv4 } from "uuid";
 import { BuildConfig } from "../../build/buildConfig";
+import { SSMParamWriter } from "../../constructs/ssm-param-writer";
+import * as importAccountAssignmentsSMJSON from "../../state-machines/import-account-assignments-asl.json";
+import * as importPermissionSetsSMJSON from "../../state-machines/import-permission-sets-asl.json";
+import * as importCurrentConfigSMJSON from "../../state-machines/import-current-config-asl.json";
 import { CrossAccountRole } from "../../constructs/cross-account-role";
-import { SSOObjectsDiscoveryPart1 } from "../../constructs/sso-objects-discovery-part1";
+import {
+  CfnQueryDefinition,
+  LogGroup,
+  RetentionDays,
+} from "aws-cdk-lib/aws-logs";
 
 function name(buildConfig: BuildConfig, resourcename: string): string {
   return buildConfig.Environment + "-" + resourcename;
 }
-
-const generateRandomString = uuidv4().toString().split("-")[0];
-const requestId = uuidv4().toString();
 
 export class SSOImportArtefactsPart1 extends Stack {
   constructor(
@@ -25,114 +39,391 @@ export class SSOImportArtefactsPart1 extends Stack {
   ) {
     super(scope, id, props);
 
-    /** Static name for temporary DynamoDB table */
-    const temporaryPermissionSetTableName = `${buildConfig.Environment}-temp-PermissionSets`;
-
-    /**
-     * We deploy SSO import artefacts part 1 for both nightly run and import
-     * current config use cases
-     */
-
-    const deploySSOImportArtefactsPart1 = new SSOObjectsDiscoveryPart1(
+    // KMS Key for importTopis
+    const ssoArtefactsKeyforImport = new Key(
       this,
-      name(buildConfig, "deploySSOImportArtefactsforNightlyRun"),
-      buildConfig
+      name(buildConfig, "ssoArtefactsKeyforImport"),
+      {
+        enableKeyRotation: true,
+        alias: name(buildConfig, "ssoArtefactsKeyforImport"),
+      }
     );
-    /**
-     * Create all deterministic resource names for importing and using it as a
-     * target for the nightly run event bridge rule
-     */
+    ssoArtefactsKeyforImport.grantEncryptDecrypt(
+      new ServicePrincipal("states.amazonaws.com")
+    );
 
-    const importCurrentConfigSMArn = `arn:aws:states:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:stateMachine:${buildConfig.Environment}-importCurrentConfigSM`;
-    const importAccountAssignmentSMArn = `arn:aws:states:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:stateMachine:${buildConfig.Environment}-importAccountAssignmentSM`;
-    const importPermissionSetSMArn = `arn:aws:states:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:stateMachine:${buildConfig.Environment}-importPermissionSetSM`;
+    // Topics that would allow cross account handlers to process import of existing configuration
+    const permissionSetImportTopic = new Topic(
+      this,
+      name(buildConfig, "permissionSetImportTopic"),
+      {
+        masterKey: ssoArtefactsKeyforImport,
+        displayName: name(buildConfig, "permissionSetImportTopic"),
+      }
+    );
+    permissionSetImportTopic.addToResourcePolicy(
+      new PolicyStatement({
+        principals: [
+          new AccountPrincipal(buildConfig.PipelineSettings.TargetAccountId),
+        ],
+        actions: ["SNS:Subscribe", "SNS:Receive"],
+        resources: [permissionSetImportTopic.topicArn],
+      })
+    );
+    new SSMParamWriter(
+      this,
+      name(buildConfig, "permissionSetImportTopicArn"),
+      buildConfig,
+      {
+        ParamNameKey: "permissionSetImportTopicArn",
+        ParamValue: permissionSetImportTopic.topicArn,
+        ReaderAccountId: buildConfig.PipelineSettings.TargetAccountId,
+      }
+    );
 
-    /** Use case specific additions below */
+    const accountAssignmentImportTopic = new Topic(
+      this,
+      name(buildConfig, "accountAssignmentImportTopic"),
+      {
+        masterKey: ssoArtefactsKeyforImport,
+        displayName: name(buildConfig, "accountAssignmentImportTopic"),
+      }
+    );
+    accountAssignmentImportTopic.addToResourcePolicy(
+      new PolicyStatement({
+        principals: [
+          new AccountPrincipal(buildConfig.PipelineSettings.TargetAccountId),
+        ],
+        actions: ["SNS:Subscribe", "SNS:Receive"],
+        resources: [accountAssignmentImportTopic.topicArn],
+      })
+    );
+    new SSMParamWriter(
+      this,
+      name(buildConfig, "accountAssignmentImportTopicArn"),
+      buildConfig,
+      {
+        ParamNameKey: "accountAssignmentImportTopicArn",
+        ParamValue: accountAssignmentImportTopic.topicArn,
+        ReaderAccountId: buildConfig.PipelineSettings.TargetAccountId,
+      }
+    );
 
-    /**
-     * Create event bridge rule for nightlyRun in a disabled state. This will be
-     * enabled if the customer chooses to have nightlyRun feature
-     */
+    /** Log group to attach to all the state machines for capturing logs */
+    const importArtefactsSMLogGroup = new LogGroup(
+      this,
+      name(buildConfig, "importArtefactsSMLogGroup"),
+      {
+        retention: RetentionDays.ONE_MONTH,
+      }
+    );
 
-    new Rule(this, name(buildConfig, "nightlyRunRule"), {
-      ruleName: name(buildConfig, "nightlyRunRule"),
-      description:
-        "Enable nightly run for aws-sso-extensions-for-enterprise governance",
-      enabled: false, // This rule is disabled on creation to accommodate race condition with the import config lambda's that are not yet created. This will be enabled as part of an additional pipeline stage after creation of the lambda functions in part2
-      schedule: Schedule.cron({ minute: "0", hour: "23" }),
-      targets: [
-        new SfnStateMachine(
-          StateMachine.fromStateMachineArn(
-            this,
-            name(buildConfig, `currentConfigSMTarget`),
-            importCurrentConfigSMArn
-          ),
-          {
-            input: RuleTargetInput.fromObject({
-              temporaryPermissionSetTableName: temporaryPermissionSetTableName,
-              importAccountAssignmentSMArn: importAccountAssignmentSMArn,
-              accountAssignmentImportTopicArn:
-                deploySSOImportArtefactsPart1.accountAssignmentImportTopic
-                  .topicArn,
-              importPermissionSetSMArn: importPermissionSetSMArn,
-              permissionSetImportTopicArn:
-                deploySSOImportArtefactsPart1.permissionSetImportTopic.topicArn,
-              PhysicalResourceId: generateRandomString,
-              eventType: "Create",
-              triggerSource: "EventBridge",
-              requestId: requestId,
-              waitSeconds: 2,
-              pageSize: 5,
-            }),
-          }
-        ),
-      ],
+    // State Machine 01 - Import Account Assignments
+
+    const importAccountAssignmentsSMRole = new Role(
+      this,
+      name(buildConfig, "importAccountAssignmentsSMRole"),
+      {
+        roleName: name(buildConfig, "importAccountAssignmentsSMRole"),
+        assumedBy: new ServicePrincipal("states.amazonaws.com"),
+      }
+    );
+    importAccountAssignmentsSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        resources: ["*"],
+        actions: ["sso:ListAccountAssignments"],
+      })
+    );
+    accountAssignmentImportTopic.grantPublish(importAccountAssignmentsSMRole);
+    ssoArtefactsKeyforImport.grantEncryptDecrypt(
+      importAccountAssignmentsSMRole
+    );
+    importAccountAssignmentsSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: ["states:DescribeExecution", "states:StopExecution"],
+        resources: ["*"],
+      })
+    );
+    importAccountAssignmentsSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: ["events:PutTargets", "events:PutRule", "events:DescribeRule"],
+        resources: [
+          `arn:aws:events:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:rule/StepFunctionsGetEventsForStepFunctionsExecutionRule`,
+        ],
+      })
+    );
+
+    importAccountAssignmentsSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        resources: ["*"],
+        actions: [
+          "identitystore:ListGroups",
+          "identitystore:ListUsers",
+          "identitystore:DescribeGroup",
+          "identitystore:DescribeUser",
+        ],
+      })
+    );
+    importAccountAssignmentsSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: [
+          "logs:CreateLogDelivery",
+          "logs:GetLogDelivery",
+          "logs:UpdateLogDelivery",
+          "logs:DeleteLogDelivery",
+          "logs:ListLogDeliveries",
+          "logs:PutResourcePolicy",
+          "logs:DescribeResourcePolicies",
+          "logs:DescribeLogGroups",
+        ],
+        resources: ["*"],
+      })
+    );
+
+    const importAccountAssignmentSM = new CfnStateMachine(
+      this,
+      name(buildConfig, "importAccountAssignmentSM"),
+      {
+        roleArn: importAccountAssignmentsSMRole.roleArn,
+        definitionString: JSON.stringify(importAccountAssignmentsSMJSON),
+        stateMachineName: name(buildConfig, "importAccountAssignmentSM"),
+        loggingConfiguration: {
+          destinations: [
+            {
+              cloudWatchLogsLogGroup: {
+                logGroupArn: importArtefactsSMLogGroup.logGroupArn,
+              },
+            },
+          ],
+          includeExecutionData: true,
+          level: "ALL",
+        },
+      }
+    );
+    importAccountAssignmentSM.node.addDependency(
+      importAccountAssignmentsSMRole
+    );
+
+    // State Machine 02 - Import Permission Sets
+
+    const importPermissionSetSMRole = new Role(
+      this,
+      name(buildConfig, "importPermissionSetSMRole"),
+      {
+        roleName: name(buildConfig, "importPermissionSetSMRole"),
+        assumedBy: new ServicePrincipal("states.amazonaws.com"),
+      }
+    );
+    importPermissionSetSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        resources: ["*"],
+        actions: [
+          "sso:ListPermissionSets",
+          "sso:DescribePermissionSet",
+          "sso:ListManagedPoliciesInPermissionSet",
+          "sso:GetInlinePolicyForPermissionSet",
+          "sso:ListTagsForResource",
+          "sso:ListAccountsForProvisionedPermissionSet",
+        ],
+      })
+    );
+    importPermissionSetSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: [
+          "dynamodb:PutItem",
+          "dynamodb:UpdateItem",
+          "dynamodb:GetItem",
+          "dynamodb:DeleteItem",
+        ],
+        resources: [
+          `arn:aws:dynamodb:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:table/${buildConfig.Environment}-temp-PermissionSets`,
+          `arn:aws:dynamodb:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:table/${buildConfig.Environment}-temp-PermissionSets/index/*`,
+        ],
+      })
+    );
+    importPermissionSetSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: ["states:StartExecution"],
+        resources: [importAccountAssignmentSM.ref],
+      })
+    );
+    importPermissionSetSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: ["states:DescribeExecution", "states:StopExecution"],
+        resources: ["*"],
+      })
+    );
+
+    importPermissionSetSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: ["events:PutTargets", "events:PutRule", "events:DescribeRule"],
+        resources: [
+          `arn:aws:events:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:rule/StepFunctionsGetEventsForStepFunctionsExecutionRule`,
+        ],
+      })
+    );
+    importPermissionSetSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: [
+          "logs:CreateLogDelivery",
+          "logs:GetLogDelivery",
+          "logs:UpdateLogDelivery",
+          "logs:DeleteLogDelivery",
+          "logs:ListLogDeliveries",
+          "logs:PutResourcePolicy",
+          "logs:DescribeResourcePolicies",
+          "logs:DescribeLogGroups",
+        ],
+        resources: ["*"],
+      })
+    );
+
+    permissionSetImportTopic.grantPublish(importPermissionSetSMRole);
+    ssoArtefactsKeyforImport.grantEncryptDecrypt(importPermissionSetSMRole);
+
+    const importPermissionSetSM = new CfnStateMachine(
+      this,
+      name(buildConfig, "importPermissionSetSM"),
+      {
+        roleArn: importPermissionSetSMRole.roleArn,
+        definitionString: JSON.stringify(importPermissionSetsSMJSON),
+        stateMachineName: name(buildConfig, "importPermissionSetSM"),
+        loggingConfiguration: {
+          destinations: [
+            {
+              cloudWatchLogsLogGroup: {
+                logGroupArn: importArtefactsSMLogGroup.logGroupArn,
+              },
+            },
+          ],
+          includeExecutionData: true,
+          level: "ALL",
+        },
+      }
+    );
+    importPermissionSetSM.node.addDependency(importPermissionSetSMRole);
+
+    // State Machine 03 - Parent state machine that orchestrates state machine 01 and 02
+
+    const importCurrentConfigSMRole = new Role(
+      this,
+      name(buildConfig, "importCurrentConfigSMRole"),
+      {
+        roleName: name(buildConfig, "importCurrentConfigSMRole"),
+        assumedBy: new ServicePrincipal("states.amazonaws.com"),
+      }
+    );
+    importCurrentConfigSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: [
+          "dynamodb:CreateTable",
+          "dynamodb:DeleteTable",
+          "dynamodb:DescribeTable",
+        ],
+        resources: [
+          `arn:aws:dynamodb:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:table/${buildConfig.Environment}-temp-PermissionSets`,
+          `arn:aws:dynamodb:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:table/${buildConfig.Environment}-temp-PermissionSets/index/*`,
+        ],
+      })
+    );
+    importCurrentConfigSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        resources: ["*"],
+        actions: ["sso:ListInstances"],
+      })
+    );
+    importCurrentConfigSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: ["states:StartExecution"],
+        resources: [importPermissionSetSM.ref],
+      })
+    );
+    importCurrentConfigSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: ["states:DescribeExecution", "states:StopExecution"],
+        resources: ["*"],
+      })
+    );
+
+    importCurrentConfigSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: ["events:PutTargets", "events:PutRule", "events:DescribeRule"],
+        resources: [
+          `arn:aws:events:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:rule/StepFunctionsGetEventsForStepFunctionsExecutionRule`,
+        ],
+      })
+    );
+    importCurrentConfigSMRole.addToPrincipalPolicy(
+      new PolicyStatement({
+        actions: [
+          "logs:CreateLogDelivery",
+          "logs:GetLogDelivery",
+          "logs:UpdateLogDelivery",
+          "logs:DeleteLogDelivery",
+          "logs:ListLogDeliveries",
+          "logs:PutResourcePolicy",
+          "logs:DescribeResourcePolicies",
+          "logs:DescribeLogGroups",
+        ],
+        resources: ["*"],
+      })
+    );
+
+    const importCurrentConfigSM = new CfnStateMachine(
+      this,
+      name(buildConfig, "importCurrentConfigSM"),
+      {
+        roleArn: importCurrentConfigSMRole.roleArn,
+        definitionString: JSON.stringify(importCurrentConfigSMJSON),
+        stateMachineName: name(buildConfig, "importCurrentConfigSM"),
+        loggingConfiguration: {
+          destinations: [
+            {
+              cloudWatchLogsLogGroup: {
+                logGroupArn: importArtefactsSMLogGroup.logGroupArn,
+              },
+            },
+          ],
+          includeExecutionData: true,
+          level: "ALL",
+        },
+      }
+    );
+
+    importCurrentConfigSM.node.addDependency(importCurrentConfigSMRole);
+
+    new CrossAccountRole(this, name(buildConfig, "ssoListRole"), buildConfig, {
+      assumeAccountID: buildConfig.PipelineSettings.TargetAccountId,
+      roleNameKey: "ssoList-ssoapi",
+      policyStatement: new PolicyStatement({
+        resources: [importCurrentConfigSM.ref],
+        actions: ["states:StartExecution"],
+      }),
     });
 
-    /** Create Cross account role to be used by nightlyRun processing lambda */
     new CrossAccountRole(
       this,
-      name(buildConfig, "nightlyRun-ssoapi-Role"),
+      name(buildConfig, "smDescribeRole"),
       buildConfig,
       {
         assumeAccountID: buildConfig.PipelineSettings.TargetAccountId,
-        roleNameKey: "nightlyRun-ssoapi",
+        roleNameKey: "smDescribe-ssoapi",
         policyStatement: new PolicyStatement({
           resources: ["*"],
-          actions: [
-            "sso:DeleteInlinePolicyFromPermissionSet",
-            "sso:DetachManagedPolicyFromPermissionSet",
-            "sso:DeletePermissionSet",
-            "sso:CreatePermissionSet",
-            "sso:GetPermissionSet",
-            "sso:TagResource",
-            "sso:ListTagsForResource",
-            "sso:ListPermissionSetProvisioningStatus",
-            "sso:UntagResource",
-            "sso:ListInstances",
-            "sso:DescribePermissionSet",
-            "sso:ProvisionPermissionSet",
-            "sso:ListPermissionSets",
-            "sso:ListAccountsForProvisionedPermissionSet",
-            "sso:PutInlinePolicyToPermissionSet",
-            "sso:AttachManagedPolicyToPermissionSet",
-            "sso:UpdatePermissionSet",
-            "sso:GetInlinePolicyForPermissionSet",
-            "sso:ListManagedPoliciesInPermissionSet",
-            "sso:DescribePermissionSetProvisioningStatus",
-            "sso:CreateAccountAssignment",
-            "sso:DescribeAccountAssignmentDeletionStatus",
-            "sso:ListAccountAssignmentDeletionStatus",
-            "sso:DescribeAccountAssignmentCreationStatus",
-            "sso:DeleteAccountAssignment",
-            "sso:ListAccountAssignmentCreationStatus",
-            "sso:ListAccountAssignments",
-            "sso:ListInstances",
-            "sso:DescribeAccountAssignmentDeletionStatus",
-            "sso:DescribePermissionSetProvisioningStatus",
-            "sso:DescribeAccountAssignmentCreationStatus",
-          ],
+          actions: ["states:DescribeExecution"],
         }),
+      }
+    );
+
+    /** CloudWatch insights query to debug errors, if any */
+    new CfnQueryDefinition(
+      this,
+      name(buildConfig, "importCurrentSSOConfiguration-errors"),
+      {
+        name: name(buildConfig, "importCurrentSSOConfiguration-errors"),
+        queryString:
+          "filter @message like 'solutionError' and details.name not like 'Catchall'| sort id asc",
+        logGroupNames: [importArtefactsSMLogGroup.logGroupName],
       }
     );
   }
