+++ lib/stacks/region-switch/aws-sso-extensions-region-switch-discover.ts
@@ -51,7 +51,7 @@ export class AwsSsoExtensionsRegionSwitchDiscover extends Stack {
       code: Code.fromAsset(
         join(__dirname, "../../", "lambda-layers", "nodejs-layer")
       ),
-      compatibleRuntimes: [Runtime.NODEJS_16_X],
+      compatibleRuntimes: [Runtime.NODEJS_14_X],
     });
 
     /**
@@ -443,7 +443,7 @@ export class AwsSsoExtensionsRegionSwitchDiscover extends Stack {
       this,
       fullname(`rsImportPermissionSetsHandler`),
       {
-        runtime: Runtime.NODEJS_16_X,
+        runtime: Runtime.NODEJS_14_X,
         layers: [rsNodeJsLayer],
         entry: join(
           __dirname,
@@ -485,7 +485,7 @@ export class AwsSsoExtensionsRegionSwitchDiscover extends Stack {
       this,
       fullname(`rsImportAccountAssignmentHandler`),
       {
-        runtime: Runtime.NODEJS_16_X,
+        runtime: Runtime.NODEJS_14_X,
         layers: [rsNodeJsLayer],
         entry: join(
           __dirname,
@@ -538,7 +538,7 @@ export class AwsSsoExtensionsRegionSwitchDiscover extends Stack {
       this,
       fullname(`updateCustomResourceHandler`),
       {
-        runtime: Runtime.NODEJS_16_X,
+        runtime: Runtime.NODEJS_14_X,
         layers: [rsNodeJsLayer],
         entry: join(
           __dirname,
@@ -576,7 +576,7 @@ export class AwsSsoExtensionsRegionSwitchDiscover extends Stack {
       this,
       fullname(`parentSMInvokeFunction`),
       {
-        runtime: Runtime.NODEJS_16_X,
+        runtime: Runtime.NODEJS_14_X,
         functionName: fullname(`parentSMInvokeFunction`),
         layers: [rsNodeJsLayer],
         entry: join(
