+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "aws-sso-extensions-for-enterprise",
-  "version": "3.2.0",
+  "version": "3.1.4",
   "bin": {
     "aws-sso-extensions-for-enterprise": "bin/aws-sso-extensions-for-enterprise.js"
   },
@@ -28,26 +28,26 @@
     "@types/aws-lambda": "^8.10.97",
     "@types/jest": "^27.5.1",
     "@types/js-yaml": "^4.0.5",
-    "@types/node": "^17.0.36",
+    "@types/node": "^17.0.34",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.26.0",
-    "@typescript-eslint/parser": "^5.26.0",
-    "aws-cdk": "^2.25.0",
-    "esbuild": "^0.14.42",
-    "eslint": "^8.16.0",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "aws-cdk": "^2.24.1",
+    "esbuild": "^0.14.39",
+    "eslint": "^8.15.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-security": "^1.5.0",
     "jest": "^28.1.0",
     "prettier": "^2.6.2",
     "prettier-plugin-jsdoc": "^0.3.38",
-    "ts-jest": "^28.0.3",
-    "ts-node": "^10.8.0",
-    "typescript": "^4.7.2",
+    "ts-jest": "^28.0.2",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.4",
     "uuid": "^8.3.2"
   },
   "dependencies": {
-    "aws-cdk-lib": "^2.25.0",
-    "constructs": "^10.1.21",
+    "aws-cdk-lib": "^2.24.1",
+    "constructs": "^10.1.10",
     "js-yaml": "^4.1.0",
     "source-map-support": "^0.5.21"
   }
