+++ test/aws-sso-extensions-for-enterprise.ts
@@ -152,15 +152,6 @@ test("Empty Stack", () => {
           unparsedEnv["Parameters"],
           "UpgradeFromVersionLessThanV303"
         ),
-        EnableNightlyRun: ensureBoolean(
-          unparsedEnv["Parameters"],
-          "EnableNightlyRun"
-        ),
-        NightlyRunRemediationMode: ensureValidString(
-          unparsedEnv["Parameters"],
-          "NightlyRunRemediationMode",
-          ["NOTIFY", "AUTOREMEDIATE"]
-        ),
       },
     };
   }
